feature: Decompose
