+++ exclude_projects.tpl
