--- gl_query/config.cfg.tpl
@@ -1,4 +1,0 @@
-[config]
-# Must include "/api/v4/"
-url = https://<GITLAB_HOST>/api/v4/
-token = YOUR_GITLAB_PERSONAL_ACCESS_TOKEN
