+++ gl_query/config.py
@@ -1,0 +1,17 @@
+### SECRET VARIABLES (configure in config.cfg)
+GITLAB_URL = str()
+HEADERS = str()
+GL_BDAY = str()
+SCAN_TYPES = list()
+LANGS = list()
+PER_PAGE = 100
+
+# COUNTER VARIABLES
+TOTAL_PROJECTS = 0
+TOTAL_PIPELINES = 0
+TOTAL_SCANS = 0
+API_REQUESTS = 0
+
+# OPTIONS VARIABLES
+HAS_PIPELINE = bool()
+
