+++ tests/test_commands.txt
@@ -1,7 +1,9 @@
-gl-query get projects -l scala --date-after DATE --csv --filename test.csv -p
-gl-query get project --project-id 3430
-gl-query get pipeline --project-id 3430 --pipeline-id 247249
-gl-query get pipelines --project-id 4748
-gl-query get scans --date-after DATE --scan-type "sast,srcclear,dependency-scan,srcclr" --all-languages
-gl-query get scans --date-after YESTERDAY  --date-before DATE --scan-type srcclr -l scala
-gl-query --version
+python3 gl-query.py get projects -l scala --date-after DATE --csv --filename test.csv -p
+python3 gl-query.py get project --project-id 3430
+python3 gl-query.py get pipeline --project-id 3430 --pipeline-id 247249
+python3 gl-query.py get pipelines --project-id 4748
+python3 gl-query.py get scans --date-after DATE --scan-type "sast,srcclear,dependency-scan,srcclr" --all-languages
+python3 gl-query.py get scans --date-after YESTERDAY  --date-before DATE --scan-type srcclr -l scala
+python3 gl-query.py get projects --has-srcclr --date-after DATE
+python3 gl-query.py get projects --no-srcclr --date-after DATE
+python3 gl-query.py get projects --search nexus
